Move mvsData/Image.hpp to image module
