+++ src/aliceVision/image/LegacyImage.hpp
