+++ src/aliceVision/mvsData/imageIO.cpp
@@ -8,8 +8,8 @@
 
 #include <aliceVision/system/Logger.hpp>
 #include <aliceVision/image/Color.hpp>
+#include <aliceVision/image/LegacyImage.hpp>
 #include <aliceVision/mvsData/Rgb.hpp>
-#include <aliceVision/mvsData/Image.hpp>
 #include <aliceVision/image/imageAlgo.hpp>
 
 #include <OpenImageIO/imageio.h>
