+++ src/aliceVision/mvsUtils/ImagesCache.hpp
@@ -7,12 +7,12 @@
 #pragma once
 
 #include <aliceVision/image/Color.hpp>
+#include <aliceVision/image/LegacyImage.hpp>
 #include <aliceVision/mvsData/Point2d.hpp>
 #include <aliceVision/mvsData/Rgb.hpp>
 #include <aliceVision/mvsData/StaticVector.hpp>
 #include <aliceVision/mvsUtils/MultiViewParams.hpp>
 #include <aliceVision/mvsData/imageIO.hpp>
-#include <aliceVision/mvsData/Image.hpp>
 
 #include <future>
 #include <mutex>
